refactor(capabilities): migrate the audio cap to the struct-hosted actor macro
Apply the ADR-0123 struct-hosted transform to the audio cap (#2377).
